chore: move classes to their own file
