+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 1. [#65](https://github.com/influxdata/influxdb-client-php/pull/65): `FluxRecord` uses descriptive message for column that doesn't exists
+1. [#68](https://github.com/influxdata/influxdb-client-php/pull/68): Moved classes to their own file
 
 ### Documentation
 1. [#65](https://github.com/influxdata/influxdb-client-php/pull/65): Documentation for the client is located at GitHub: https://influxdata.github.io/influxdb-client-php/
