+++ src/InfluxDB2/FluxCsvParser.php
@@ -303,7 +303,3 @@ class FluxCsvParser
         $this->response->close();
     }
 }
-
-class FluxQueryError extends RuntimeException
-{
-}
