+++ src/InfluxDB2/FluxQueryError.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace InfluxDB2;
+
+use RuntimeException;
+
+class FluxQueryError extends RuntimeException
+{
+}
