+++ src/InfluxDB2/WriteApi.php
@@ -204,40 +204,3 @@ class WriteApi extends DefaultApi implements Writer
         return isset($precision) ? $precision : $this->options["$optionName"];
     }
 }
-
-/**
- * Item for batching queue
- */
-class BatchItem
-{
-    /** @var BatchItemKey */
-    public $key;
-    /** @var string */
-    public $data;
-
-    public function __construct($key, $data)
-    {
-        $this->key = $key;
-        $this->data = $data;
-    }
-}
-
-/**
- * Key for batch item
- */
-class BatchItemKey
-{
-    /** @var string */
-    public $bucket;
-    /** @var string */
-    public $org;
-    /** @var WritePrecision */
-    public $precision;
-
-    public function __construct($bucket, $org, $precision)
-    {
-        $this->bucket = $bucket;
-        $this->org = $org;
-        $this->precision = $precision;
-    }
-}
